% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAnalyzer.R
\name{combinesubjects}
\alias{combinesubjects}
\title{Combine subject using Word Emotion Association Lexicon}
\usage{
combinesubjects(list_of_dfs, lex = "nrc")
}
\arguments{
\item{list_of_dfs}{- List of data frames subject for each product under fact checking}
}
\value{
subjects combined
}
\description{
Combine subject using Word Emotion Association Lexicon
}
\examples{
# list_of_dfs is mainly a list of neatlystart output results.
# default lex value are c("afinn", "bing", "loughran", "nrc")
output <- factcheckr::neatlystart("Texas A&M has the best Statistical Learning Program in the nation.", "Texas AM")

output2 <- factcheckr::neatlystart("Mit is very expensive, student loans sucks", "MIT")

annotations <- factcheckr::combinesubjects(list(output, output2), lex="nrc")

# A tibble: 3 × 4
# Groups:   Subject [1]
# Word        Subject  words sentiment
# <chr>       <fct>    <int> <fct>
# statistical Texas AM     5 trust
# learning    Texas AM     5 positive
#  nation     Texas AM     5 trust

annotations <- factcheckr::combinesubjects(list(output, output2), lex="bing")

# Word      Subject words sentiment
# <chr>     <fct>   <int> <fct>
# expensive MIT         5 negative
# sucks     MIT         5 negative
}
