% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestData.R
\name{createTestData}
\alias{createTestData}
\title{Basic helper function to rapidly work with the example Hotel_Reviews
dataset shipped with this package. It collapses reviews into a single text}
\usage{
createTestData(dataset, hotelName = "Red Roof Inn Cedar Rapids")
}
\arguments{
\item{dataset}{- the hotel review dataset}

\item{hotelName}{- the hotel name whose reviews should be extracted}
}
\value{
the text collapse into a single text
}
\description{
Basic helper function to rapidly work with the example Hotel_Reviews
dataset shipped with this package. It collapses reviews into a single text
}
\examples{

createTestData(dataset=Hotel_Reviews, hotelName="Red Roof Inn Cedar Rapids")


}
