% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{.polarityPlot}
\alias{.polarityPlot}
\title{Polarity visualization for each subject. This is the ratio of the number of positive
emotion words divided by the number of negative words}
\usage{
.polarityPlot(subjects)
}
\arguments{
\item{subjects}{- subject under fact checking}
}
\value{
the plot
}
\description{
Polarity visualization for each subject. This is the ratio of the number of positive
emotion words divided by the number of negative words
}
\examples{

.polarityPlot(subjects=polarityChanges(nrcSubjects))

}
