% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{.toptermplot}
\alias{.toptermplot}
\title{Helper function to visualize the top n words for the core emotion categories.}
\usage{
.toptermplot(topwords, show_legend = FALSE)
}
\arguments{
\item{topwords}{- tibble of top words}

\item{show_legend}{- boolean whether legend should be display in the plot or not}
}
\value{
tibble of top words by emotion
}
\description{
Helper function to visualize the top n words for the core emotion categories.
}
\examples{

toptermplot(topwords=.topterms(nrcSubjects))

}
