% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAnalyzer.R
\name{emotionFrequency}
\alias{emotionFrequency}
\title{Summarizes the results of the Sentiment Analysis and calculate the percentages of the
prevalence of emotions across different subjects(Products) under analysis.}
\usage{
emotionFrequency(subjectsAnnotations)
}
\arguments{
\item{subjectsAnnotations}{- the annotation of different emotion by subject in this corpus}
}
\value{
emotion emotion frequency
}
\description{
Summarizes the results of the Sentiment Analysis and calculate the percentages of the
prevalence of emotions across different subjects(Products) under analysis.
}
\examples{

 out <- emotionFrequency(subjectsAnnotations)

# Subject  sentiment sentiment_freq words percentage
# <fct>    <fct>              <int> <int>      <dbl>
#  Texas AM positive             1     5         20
#  Texas AM trust                2     5         40
}
\seealso{
\code{\link{combineSubjects}}, \code{\link{emotionFrequency}}, \code{\link{topterms}}
}
