% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fact_check_insights}
\alias{fact_check_insights}
\title{fact_check_insights}
\format{
A data frame with 74232 rows and 4 columns:
\describe{
\item{date_published}{Date of publication of the claim (Date)}
\item{author_name}{Author Name of the reviewer (factor)}
\item{review_rating_type}{Review rating type (factor)}
\item{n}{Number of claims (integer)}
}
}
\source{
Fact Check Insights
}
\usage{
fact_check_insights
}
\description{
Count of Claims Reviewed in the Fact Check Insights Dataset
}
\keyword{datasets}
