% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{frequencyPlot}
\alias{frequencyPlot}
\title{Plot top Words frequency under the corpus}
\usage{
frequencyPlot(x, top = 10)
}
\arguments{
\item{x}{- tibble of Word and subject columns created from neatlyStart function}

\item{top}{- the number of word to show in the frequency plot}
}
\value{
the plot
}
\description{
Plot top Words frequency under the corpus
}
\examples{
corpus <- "Comfort Suites hotels in the Bahamas is dope. check it out"
frequencyPlot(x=neatlyStart(corpus=corpus, subject="Atlantis"), top = 100)

}
