% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hotel_Reviews}
\alias{Hotel_Reviews}
\title{Hotel_Reviews}
\format{
A data frame with 35912 rows and 19 columns.
}
\source{
https://www.kaggle.com/datasets/datafiniti/hotel-reviews.
}
\description{
The  hotel reviews for Sentiment Analysis
}
\details{
The Hotel_Reviews dataset

This dataset 782 hotels names with reviews
}
\examples{
data(Hotel_Reviews)
}
