% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAnalyzer.R
\name{loadlexicon}
\alias{loadlexicon}
\title{Load the correct lexicon}
\usage{
loadlexicon(lexicon = "nrc")
}
\arguments{
\item{lexicon}{-  lexicon type to find}
}
\value{
the lexicon
}
\description{
Load the correct lexicon
}
\examples{

loadlexicon(lexicon ="nrc")

}
