% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAnalyzer.R
\name{polarityChange}
\alias{polarityChange}
\title{An alternative approach to monitoring shifts in polarity across time involves
computing rolling or moving averages. It is important to recognize, that while
rolling averages are not the most effective means for monitoring temporal changes,
they serve as a technique for smoothing out erratic time-series data.}
\usage{
polarityChange(subjectsAnnotations)
}
\arguments{
\item{subjectsAnnotations}{- subject annotation to find polarity changes over time}
}
\value{
the polarity changes over time.
}
\description{
An alternative approach to monitoring shifts in polarity across time involves
computing rolling or moving averages. It is important to recognize, that while
rolling averages are not the most effective means for monitoring temporal changes,
they serve as a technique for smoothing out erratic time-series data.
}
\examples{

 nsObamaSpeech <- neatlyStart(corpus=ObamaVictorySpeech, subject="Obama Speech")
 nrcSubjects <- combineSubjects(list(nsObamaSpeech), lex="nrc")

 out <- polarityChange(subjectsAnnotations = nrcSubjects)

# Groups:   Subject, sentiment [1]
# Subject, id ,  rmean
# <fct>   <int>  <lgl>
}
\seealso{
\code{\link{combineSubjects}}, \code{\link{emotionFrequency}},
         \code{\link{topterms}}, \code{\link{polarityChange}}
}
