% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{removeStopwords}
\alias{removeStopwords}
\title{Remove additional stopwords under fact checking}
\usage{
removeStopwords(words_dictionary, stopwords = data.frame(word = c()))
}
\arguments{
\item{words_dictionary}{-  dictionary of words array.}

\item{stopwords}{- The data frame of words to remove.}
}
\value{
dictionary of words with no stop words.
}
\description{
Remove additional stopwords under fact checking
}
\examples{
words <- neatlyStart(corpus="Texas A&M has the best Statistical Learning
            Program in the nation.", subject="University")
Output <- removeStopwords(words, data.frame(word = c("texas")))

# A tibble: 5 × 2
# Word          Subject
# <chr>         <chr>
# 1 statistical University
# 2 learning    University
# 3 program     University
# 4 nation      University
#
# Texas has been removed from the texas as it could be words that means nothing in the study
}
\seealso{
\code{\link{trimTextl}}, \code{\link{trimTextr}}, \code{\link{strtokwords}}

\code{\link{strtoksentence}}, \code{\link{emotionFrequency}}, \code{\link{topterms}}

\code{\link{neatlyStart}}, \code{\link{removeStopwords}}, \code{\link{combineSubjects}}
}
