% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{sentimentPlot}
\alias{sentimentPlot}
\title{Sentiment score visualization base on words score. It plots the scores along
the axis labeled with both positive as well as negative words.}
\usage{
sentimentPlot(x, lexicon = "bing", cutoffScore = 100)
}
\arguments{
\item{x}{- A vector of words.}

\item{lexicon}{- one of the lexicon (default is bing.)}

\item{cutoffScore}{- cut off score to include in the plot}
}
\value{
Sentiment Score plot
}
\description{
Sentiment score visualization base on words score. It plots the scores along
the axis labeled with both positive as well as negative words.
}
\examples{

sentimentPlot(x=neatlyStart(corpus=ObamaVictorySpeech, subject="Comfort Suites"))

}
