% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{strtoksentence}
\alias{strtoksentence}
\title{Extracts sentences from text if there is any.}
\usage{
strtoksentence(text)
}
\arguments{
\item{text}{- text to be tokenized into sentences}
}
\value{
an arrays of tokens sentences to return
}
\description{
Extracts sentences from text if there is any.
}
\examples{
Output <- strtoksentence("This is a very long character vector. Why is it so long? I think lng. is short for long")
# Output[1]: "This is a very long character vector"
# Output[2]: "Why is it so long?"
# Output[3]: "I think lng. is short for long"
}
