% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{strtokwords}
\alias{strtokwords}
\title{Turn a text into words tokens}
\usage{
strtokwords(text)
}
\arguments{
\item{text}{-text to be tokenized}
}
\value{
an arrays of words tokens
}
\description{
Turn a text into words tokens
}
\examples{
Output <- strtokwords("This is a very long character vector.")
# Output[1]: "This"
# Output[2]: "is"
# Output[3]: "a"
# Output[4]: "very"
# Output[5]: "long"
# Output[6]: "character"
# Output[7]: "vector"
}
\seealso{
\code{\link{trimTextl}}, \code{\link{trimTextr}}, \code{\link{strtokwords}}

\code{\link{strtoksentence}}, \code{\link{emotionFrequency}}, \code{\link{topterms}}

\code{\link{neatlystart}}, \code{\link{removeStopwords}}, \code{\link{combinesubjects}}
}
