% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{trimText}
\alias{trimText}
\title{Clean text before analysis.
This function takes a text as input and returns a new text with all extra spaces removed}
\usage{
trimText(str, left = FALSE, right = FALSE)
}
\arguments{
\item{str}{- text to remove extra spaces left, middle and right}

\item{left}{- boolean flag if extra spaces should be removed from the left only}

\item{right}{- boolean flag if extra spaces should be removed from the right only}
}
\value{
the text already with extra spaces removed
}
\description{
Clean text before analysis.
This function takes a text as input and returns a new text with all extra spaces removed
}
\examples{
trimText("  This   is  String  ")
# Output: "This is String"

 trimText("  This   is  String  ", left=TRUE)
# Output: "This   is  String  "

trimText("  This   is  String  ", right=TRUE)
# Output: "  This   is  String"
}
