% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{trimTextl}
\alias{trimTextl}
\title{helper function to remove left spacing in a text, necessary for data cleanup}
\usage{
trimTextl(str)
}
\arguments{
\item{str}{- the string text to remove extra spacing}
}
\value{
the text with spacing removed on the left
}
\description{
helper function to remove left spacing in a text, necessary for data cleanup
}
\examples{
trimTextl("  This   is  String  ")
# Output: "This   is  String  "
}
\seealso{
\code{\link{trimText}}, \code{\link{trimTextr}}, \code{\link{strtokwords}}

\code{\link{strtoksentence}}, \code{\link{emotionFrequency}}, \code{\link{topterms}}

\code{\link{neatlystart}}, \code{\link{removeStopwords}}, \code{\link{combinesubjects}}
}
