% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{trimTextr}
\alias{trimTextr}
\title{helper function to remove extra right spacing.}
\usage{
trimTextr(str)
}
\arguments{
\item{str}{-  text to remove extra spacing}
}
\value{
text with spacing removed on the left
}
\description{
helper function to remove extra right spacing.
}
\examples{
trimTextr("  This   is  String  ")
# Output: "  This   is  String"
}
