% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrumpSpeech.R
\docType{data}
\name{TrumpVictorySpeech}
\alias{TrumpVictorySpeech}
\title{VictorySpeech dataset}
\format{
A single text data object
}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems.
}
\description{
A famous dataset in Sentiment Analysis.
# @usage data(TrumpVictorySpeech)
}
\details{
The TrumpVictorySpeech dataset

This dataset contains President Obama and Donald Trump Victory speeches,
}
\examples{
data(TrumpVictorySpeech)
}
