% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{.wordcloudPlot}
\alias{.wordcloudPlot}
\title{internal helper function to draw wordcloud}
\usage{
.wordcloudPlot(corpus, lexicon = "nrc", maxWords = 50)
}
\arguments{
\item{corpus}{- corpus to plot wordcloud}

\item{lexicon}{- lexicon to use. see documentation for different lexicon supported}

\item{maxWords}{- max number of words to add to this word cloud}

\item{sort_default}{- sorting parameter}
}
\value{
the graph showing the wordcloud of the sentiment analysis
}
\description{
internal helper function to draw wordcloud
}
