% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{ggplot3}
\alias{ggplot3}
\title{This function will provide the most frequently use words in a text plot.
Display emotions by subject and re-level sentiment so that the different core emotions are}
\usage{
ggplot3(
  text,
  top = 10,
  emotions_by_subject = FALSE,
  graphType = c("frequency", "emotion", "topterm", "movingaverage", "polarity",
    "wordcloud")
)
}
\arguments{
\item{text}{- text to use to produce the graph, subject to draw emotion from}

\item{top}{- indicated the limit to return after the operation is complete}

\item{emotions_by_subject}{- Boolean for dictating emotion graph plot}

\item{graphType}{- the type of graph to plot}
}
\value{
will plot the graph type passed to the function
}
\description{
This function will provide the most frequently use words in a text plot.
Display emotions by subject and re-level sentiment so that the different core emotions are
}
\examples{

ggplot3(nrcResult, top=10, "frequency")


}
