% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextUtils.R
\name{neatlyStart}
\alias{neatlyStart}
\title{clean and convert text to lowercase, then to words, remove any extra spaces
then add new column pass as parameter in the subject, remove all stops words and finally remove any
non English words.}
\usage{
neatlyStart(corpus = "", subject = "domain")
}
\arguments{
\item{corpus}{- Text corpus under fact checking to clean before starting the analysis}

\item{subject}{- subject domain to apply this study.}
}
\value{
An Array(vector) of words in this corpus.
}
\description{
clean and convert text to lowercase, then to words, remove any extra spaces
then add new column pass as parameter in the subject, remove all stops words and finally remove any
non English words.
}
\examples{
Output <- neatlyStart(corpus="Texas A&M has the best Statistical Learning Program.",
                   subject="University")
# A tibble: 5 × 2
# Word          Subject
# <chr>         <chr>
# 1 texas       University
# 2 statistical University
# 3 learning    University
# 4 program     University
# 5 nation      University
#
# Clearly has, the, best, in are stopwords and have been removed
}
\seealso{
\code{\link{trimTextl}}, \code{\link{trimTextr}}, \code{\link{strtokwords}}

\code{\link{strtoksentence}}, \code{\link{emotionFrequency}}, \code{\link{topterms}}

\code{\link{neatlyStart}}, \code{\link{removeStopwords}}, \code{\link{combineSubjects}}
}
