% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObamaSpeech.R
\docType{data}
\name{ObamaVictorySpeech}
\alias{ObamaVictorySpeech}
\title{VictorySpeech dataset}
\format{
A single text data object
}
\source{
https://www.npr.org/2008/11/05/96624326/transcript-of-barack-obamas-victory-speech.
}
\description{
A famous dataset in Sentiment Analysis.
}
\details{
The Obama Victory Speech dataset

This dataset contains President Obama and Donald Trump Victory speeches,
}
\examples{
data(ObamaVictorySpeech)
}
