% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAnalyzer.R
\name{topterms}
\alias{topterms}
\title{Check words that have contributed to the emotionality of scores. we investigate, which words are more
important for the emotion scores within each subject. For the sake of interpretability,
we will remove several core emotion categories and also the polarity.}
\usage{
topterms(subjects_annotation, min_top_words = 4)
}
\arguments{
\item{subjects_annotation}{- the subject annotation}

\item{top_words}{- the number of top words to return if any}
}
\value{
topwords associated with the lexicon
}
\description{
Check words that have contributed to the emotionality of scores. we investigate, which words are more
important for the emotion scores within each subject. For the sake of interpretability,
we will remove several core emotion categories and also the polarity.
}
\examples{


 out <- topterms (subjects_annotation)

# Groups:   Subject, sentiment [1]
# Subject  Word        sentiment     n score
# <fct>    <chr>       <fct>     <int> <dbl>
# Texas AM nation      trust         1   0.5
# Texas AM statistical trust         1   0.5
}
\seealso{
\code{\link{combinesubjects()}}, \code{\link{emotionFrequency()}}
}
