% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrumpSpeech.R
\docType{data}
\name{TrumpVictorySpeech}
\alias{TrumpVictorySpeech}
\title{VictorySpeech dataset}
\format{
A single text data object
}
\source{
https://www.cnn.com/2016/11/09/politics/donald-trump-victory-speech/index.html.
}
\description{
A famous dataset in Sentiment Analysis.
}
\details{
The TrumpVictorySpeech dataset

This dataset contains President Obama and Donald Trump Victory speeches,
}
\examples{
data(TrumpVictorySpeech)
}
